+++ Denoiser/IntergralImageComputerTBB.cpp
@@ -8,7 +8,7 @@ namespace Denoise
 	IntergralImageComputerTBB::IntergralImageComputerTBB(const ImageBlockProcessorSettings& settings,
 		const ImageBlockProcessorSettingsInternal& settingsInternal,
 		const std::vector<std::pair<int, int> >& shifts,
-		std::vector<std::vector<std::vector<float> > >& distanceImage,
+		std::vector<std::vector<std::vector<double> > >& distanceImage,
 		std::vector<std::vector<std::vector<double> > >& integralImage,
 		const Image& image) :
 		m_settings(settings), m_settingsInternal(settingsInternal), m_shifts(shifts),
